+++ packages/melonjs/src/video/canvas/canvas_renderer.js
@@ -109,34 +109,40 @@ export default class CanvasRenderer extends Renderer {
 
 	/**
 	 * set a blend mode for the given context. <br>
-	 * Supported blend mode between Canvas and WebGL renderer : <br>
-	 * - "normal" : this is the default mode and draws new content on top of the existing content <br>
-	 * <img src="../images/normal-blendmode.png" width="510"/> <br>
-	 * - "multiply" : the pixels of the top layer are multiplied with the corresponding pixel of the bottom layer. A darker picture is the result. <br>
-	 * <img src="../images/multiply-blendmode.png" width="510"/> <br>
-	 * - "additive or lighter" : where both content overlap the color is determined by adding color values. <br>
-	 * <img src="../images/lighter-blendmode.png" width="510"/> <br>
-	 * - "screen" : The pixels are inverted, multiplied, and inverted again. A lighter picture is the result (opposite of multiply) <br>
-	 * <img src="../images/screen-blendmode.png" width="510"/> <br>
+	 * All renderers support: <br>
+	 * - "normal" : draws new content on top of the existing content <br>
+	 * - "multiply" : the pixels are multiplied, resulting in a darker picture <br>
+	 * - "add", "additive", or "lighter" : color values are added together <br>
+	 * - "screen" : pixels are inverted, multiplied, and inverted again <br>
+	 * Canvas only (WebGL2 supports darken/lighten, others fall back to normal): <br>
+	 * - "overlay", "darken", "lighten", "color-dodge", "color-burn",
+	 *   "hard-light", "soft-light", "difference", "exclusion" <br>
 	 * @see https://developer.mozilla.org/en-US/docs/Web/API/CanvasRenderingContext2D/globalCompositeOperation
-	 * @param {string} [mode="normal"] - blend mode : "normal", "multiply", "lighter, "additive", "screen"
+	 * @param {string} [mode="normal"] - blend mode
 	 * @param {CanvasRenderingContext2D} [context]
 	 */
 	setBlendMode(mode = "normal", context) {
 		context = context || this.getContext();
 		this.currentBlendMode = mode;
 		switch (mode) {
-			case "screen":
-				context.globalCompositeOperation = "screen";
-				break;
-
 			case "lighter":
 			case "additive":
+			case "add":
 				context.globalCompositeOperation = "lighter";
 				break;
 
 			case "multiply":
-				context.globalCompositeOperation = "multiply";
+			case "screen":
+			case "overlay":
+			case "darken":
+			case "lighten":
+			case "color-dodge":
+			case "color-burn":
+			case "hard-light":
+			case "soft-light":
+			case "difference":
+			case "exclusion":
+				context.globalCompositeOperation = mode;
 				break;
 
 			default: // normal