+++ packages/melonjs/src/video/webgl/webgl_renderer.js
@@ -725,17 +725,17 @@ export default class WebGLRenderer extends Renderer {
 
 	/**
 	 * set a blend mode for the given context. <br>
-	 * Supported blend mode between Canvas and WebGL renderer : <br>
-	 * - "normal" : this is the default mode and draws new content on top of the existing content <br>
-	 * <img src="../images/normal-blendmode.png" width="510"/> <br>
-	 * - "multiply" : the pixels of the top layer are multiplied with the corresponding pixel of the bottom layer. A darker picture is the result. <br>
-	 * <img src="../images/multiply-blendmode.png" width="510"/> <br>
-	 * - "additive or lighter" : where both content overlap the color is determined by adding color values. <br>
-	 * <img src="../images/lighter-blendmode.png" width="510"/> <br>
-	 * - "screen" : The pixels are inverted, multiplied, and inverted again. A lighter picture is the result (opposite of multiply) <br>
-	 * <img src="../images/screen-blendmode.png" width="510"/> <br>
-	 * @see https://developer.mozilla.org/en-US/docs/Web/API/CanvasRenderingContext2D/globalCompositeOperation
-	 * @param {string} [mode="normal"] - blend mode : "normal", "multiply", "lighter", "additive", "screen"
+	 * All renderers support: <br>
+	 * - "normal" : draws new content on top of the existing content <br>
+	 * - "multiply" : the pixels are multiplied, resulting in a darker picture <br>
+	 * - "add", "additive", or "lighter" : color values are added together <br>
+	 * - "screen" : pixels are inverted, multiplied, and inverted again <br>
+	 * WebGL2 additionally supports: <br>
+	 * - "darken" : retains the darkest pixels of both layers <br>
+	 * - "lighten" : retains the lightest pixels of both layers <br>
+	 * Other CSS blend modes (overlay, color-dodge, etc.) are only supported by the Canvas renderer
+	 * and will fall back to "normal" in WebGL. <br>
+	 * @param {string} [mode="normal"] - blend mode
 	 * @param {WebGLRenderingContext} [gl] - a WebGL context
 	 */
 	setBlendMode(mode = "normal", gl = this.gl) {
@@ -746,19 +746,46 @@ export default class WebGLRenderer extends Renderer {
 
 			switch (mode) {
 				case "screen":
+					gl.blendEquation(gl.FUNC_ADD);
 					gl.blendFunc(gl.ONE, gl.ONE_MINUS_SRC_COLOR);
 					break;
 
 				case "lighter":
 				case "additive":
+				case "add":
+					gl.blendEquation(gl.FUNC_ADD);
 					gl.blendFunc(gl.ONE, gl.ONE);
 					break;
 
 				case "multiply":
+					gl.blendEquation(gl.FUNC_ADD);
 					gl.blendFunc(gl.DST_COLOR, gl.ONE_MINUS_SRC_ALPHA);
 					break;
 
+				case "darken":
+					if (this.WebGLVersion > 1) {
+						gl.blendEquation(gl.MIN);
+						gl.blendFunc(gl.ONE, gl.ONE);
+					} else {
+						gl.blendEquation(gl.FUNC_ADD);
+						gl.blendFunc(gl.ONE, gl.ONE_MINUS_SRC_ALPHA);
+						this.currentBlendMode = "normal";
+					}
+					break;
+
+				case "lighten":
+					if (this.WebGLVersion > 1) {
+						gl.blendEquation(gl.MAX);
+						gl.blendFunc(gl.ONE, gl.ONE);
+					} else {
+						gl.blendEquation(gl.FUNC_ADD);
+						gl.blendFunc(gl.ONE, gl.ONE_MINUS_SRC_ALPHA);
+						this.currentBlendMode = "normal";
+					}
+					break;
+
 				default:
+					gl.blendEquation(gl.FUNC_ADD);
 					gl.blendFunc(gl.ONE, gl.ONE_MINUS_SRC_ALPHA);
 					this.currentBlendMode = "normal";
 					break;